chore: remove redundant codes and update deps

+++ src/models/madhab.rs
@@ -12,6 +12,7 @@ use serde::{Deserialize, Serialize};
 /// Setting for the Asr prayer time.
 /// For Hanafi madhab, the Asr is bit later
 /// than that of the Shafi madhab.
+
 #[derive(PartialEq, Eq, Debug, Copy, Clone, Hash)]
 #[cfg_attr(feature = "serde", derive(Serialize, Deserialize))]
 #[cfg_attr(feature = "schemars", derive(JsonSchema))]
@@ -21,12 +22,14 @@ pub enum Madhab {
 }
 
 impl Madhab {
-    pub fn shadow(&self) -> i32 {
-        *self as i32
+    /// Shadow length
+    pub fn shadow(&self) -> f64 {
+        *self as i32 as f64
     }
 }
 
 #[cfg(test)]
+#[allow(clippy::float_cmp)]
 mod tests {
     use super::*;
 
@@ -34,13 +37,13 @@ mod tests {
     fn shafi_shadow() {
         let shafi = Madhab::Shafi;
 
-        assert_eq!(shafi.shadow(), 1);
+        assert_eq!(shafi.shadow(), 1.);
     }
 
     #[test]
     fn hanafi_shadow() {
         let hanafi = Madhab::Hanafi;
 
-        assert_eq!(hanafi.shadow(), 2);
+        assert_eq!(hanafi.shadow(), 2.);
     }
 }
